Adding functionality to refresh the idToken and thus avoid to ask the user to login again.
This is not super well done now, as it assumes that the idToken will be
refreshed before being expired. I set the expiration date for the id
token to 100h, a period of time reasonable enough to expect the app to
be used in.

I guess the ideal would be to only store the refresh token and then
wrap each call to the api with a call to refresh the token and with the
obtained id_token, call the api. This would make sure that the id token
is always valid, but increases user-actions latency.